requirements.txt
Django==1.8.5
djangorestframework==3.2.4
requests==2.8.0
stripe==1.27.1
wheel==0.24.0
